+++ packages/go_router/lib/src/route.dart
@@ -233,6 +233,10 @@ class GoRoute {
   ///
   /// See [Sub-routes](https://github.com/flutter/packages/blob/main/packages/go_router/example/lib/sub_routes.dart)
   /// for a complete runnable example.
+  ///
+  /// If [BuildContext.dependOnInheritedWidgetOfExactType] is used during the
+  /// redirection (which is how `of` method is usually implemented), a
+  /// re-evaluation will be triggered if the [InheritedWidget] changes.
   final List<GoRoute> routes;
 
   /// An optional redirect function for this route.
@@ -295,7 +299,8 @@ class GoRoute {
   Map<String, String> extractPathParams(RegExpMatch match) =>
       extractPathParameters(_pathParams, match);
 
-  static String? _noRedirection(GoRouterState state) => null;
+  static String? _noRedirection(BuildContext context, GoRouterState state) =>
+      null;
 
   static Widget _invalidBuilder(
     BuildContext context,